Move API error types to `Cardano.Wallet.Api.Types.Error`.
